refactor: reorganize to abstract common donor code
