+++ src/encoded/static/components/browse/browse-view/BrowseDonor.js
@@ -1,11 +1,7 @@
-import React, { useState, useEffect, useMemo } from 'react';
-import * as _ from 'underscore';
+import React, { useState, useEffect } from 'react';
 import { SelectAllFilesButton } from '../../static-pages/components/SelectAllAboveTableComponent';
-import { SelectionItemCheckbox } from '@hms-dbmi-bgm/shared-portal-components/es/components/browse/components/SelectedItemsController';
 import { SearchView as CommonSearchView } from '@hms-dbmi-bgm/shared-portal-components/es/components/browse/SearchView';
 import { Schemas } from '../../util';
-import { tissueToCategory } from '../../util/data';
-import { Alerts } from '@hms-dbmi-bgm/shared-portal-components/es/components/ui/Alerts';
 import { BrowseViewControllerWithSelections } from '../../static-pages/components/TableControllerWithSelections';
 import { BrowseViewAboveFacetListComponent } from './BrowseViewAboveFacetListComponent';
 import { BrowseViewAboveSearchTableControls } from './BrowseViewAboveSearchTableControls';
@@ -14,679 +10,22 @@ import {
     BROWSE_LINKS,
     NoResultsBrowseModal,
 } from '../BrowseView';
-import { columnExtensionMap as originalColExtMap } from '../columnExtensionMap';
 import { transformedFacets } from '../SearchView';
-import { CustomTableRowToggleOpenButton } from '@hms-dbmi-bgm/shared-portal-components/es/components/browse/components/table-commons/basicColumnExtensionMap';
-import { LocalizedTime } from '@hms-dbmi-bgm/shared-portal-components/es/components/ui/LocalizedTime';
 import { BrowseDonorVizWrapper } from './BrowseDonorVizWrapper';
-import { valueTransforms } from '@hms-dbmi-bgm/shared-portal-components/es/components/util';
 import { DonorMetadataDownloadButton } from '../../shared/DonorMetadataDownloadButton';
-import { formatTissueData, formatAssayData } from './BrowseProtectedDonor';
+import {
+    customRenderDetailPane,
+    createBaseDonorColumnExtensionMap,
+} from './BrowseDonorBase';
 
-/**
- * TissueDetailPane component that displays tissue data for a donor. Fetches tissue
- * data based on the donor's display title and formats it for display.
- * @param {*} itemDetails - The details of the item to display.
- * @returns TissueDetailPane component that displays tissue data for a donor.
- */
-const TissueDetailPane = React.memo(function TissueDetailPane({
-    itemDetails,
-    panelDetails,
-}) {
-    const [tissueData, setTissueData] = useState(null);
-
-    useEffect(() => {
-        const searchURL = `/search/?type=File&${BROWSE_STATUS_FILTERS}&dataset!=No+value&donors.display_title=${itemDetails.display_title}`;
-
-        // Use cached search results if available from parent
-        if (panelDetails?.searchCache) {
-            setTissueData(
-                formatTissueData(
-                    panelDetails?.searchCache?.facets?.find(
-                        (f) => f.field === 'sample_summary.tissues'
-                    )?.terms || []
-                )
-            );
-        } else {
-            panelDetails.searchRequest(searchURL);
-        }
-    }, [panelDetails.searchCache]);
-
-    return tissueData && Object?.keys(tissueData)?.length > 0 ? (
-        <div className="detail-content">
-            <div className="detail-header">
-                <i className="icon icon-lungs fas"></i>
-                <b>
-                    {/* Calculate total tissue count */}
-                    {Object.keys(tissueData).reduce(
-                        (acc, key) => acc + tissueData[key].values.length,
-                        0
-                    )}{' '}
-                </b>
-                Tissues for Donor {itemDetails.display_title}
-            </div>
-            <div className="detail-body">
-                {Object?.keys(tissueData).map((category, i) => {
-                    const tissues = tissueData[category]['values'] || [];
-
-                    return (
-                        <div key={i} className="tissue-category">
-                            <div className="header-container">
-                                <h3>{tissueData[category]?.title}</h3>
-                            </div>
-                            <div className="tissue-list-container">
-                                {tissues.length > 0 ? (
-                                    <ul>
-                                        {tissues
-                                            .sort((a, b) => a.localeCompare(b))
-                                            .map((tissue, j) => {
-                                                // Create a link to search for files with this tissue
-                                                return (
-                                                    <li key={j}>
-                                                        <span>
-                                                            <a
-                                                                href={`/browse/?type=File&${BROWSE_STATUS_FILTERS}&dataset!=No+value&donors.display_title=${itemDetails.display_title}&sample_summary.tissues=${tissue}`}
-                                                                target="_blank"
-                                                                rel="noreferrer noopener">
-                                                                {tissue}
-                                                            </a>
-                                                        </span>
-                                                    </li>
-                                                );
-                                            })}
-                                    </ul>
-                                ) : (
-                                    <span className="text-secondary">N/A</span>
-                                )}
-                            </div>
-                        </div>
-                    );
-                })}
-            </div>
-        </div>
-    ) : (
-        <div className="detail-content">
-            <div className="detail-header">
-                <i className="icon icon-spin icon-circle-notch"></i>
-            </div>
-        </div>
-    );
-});
-
-/**
- * AssayDetailPane component that displays assay data for a donor. Fetches assay
- * data based on the donor's display title and formats it for display.
- * @param {*} itemDetails - The details of the item to display.
- * @returns AssayDetailPane component that displays tissue data for a donor.
- */
-const AssayDetailPane = React.memo(function AssayDetailPane({
-    itemDetails,
-    panelDetails,
-}) {
-    const [assayData, setAssayData] = useState(null);
-
-    useEffect(() => {
-        const searchURL = `/search/?type=File&${BROWSE_STATUS_FILTERS}&dataset!=No+value&donors.display_title=${itemDetails.display_title}`;
-
-        // Use cached search results if available from parent
-        if (panelDetails?.searchCache) {
-            const assayFacets = panelDetails?.searchCache?.facets?.find(
-                (f) => f.field === 'assays.display_title'
-            );
-
-            // Pull out terms and label overrides from facet
-            const termsFromFacets = assayFacets?.terms || [];
-            const labelOverridesFromFacets = assayFacets?.label_overrides || {};
-
-            // Get assay terms from facet
-            const assayTerms = formatAssayData(
-                termsFromFacets,
-                labelOverridesFromFacets
-            );
-
-            setAssayData(assayTerms);
-        } else {
-            panelDetails.searchRequest(searchURL);
-        }
-    }, [panelDetails.searchCache]);
-
-    return assayData && Object?.keys(assayData)?.length > 0 ? (
-        <div className="detail-content">
-            <div className="detail-header">
-                <i className="icon icon-dna fas"></i>
-                <b>
-                    {/* Calculate total tissue count */}
-                    {Object.keys(assayData).reduce((acc, key) => {
-                        return acc + assayData[key].values.length;
-                    }, 0)}{' '}
-                </b>
-                Assays across all tissues
-            </div>
-            <div className="detail-body">
-                {Object?.keys(assayData).map((category, i) => {
-                    const {
-                        title,
-                        values = [],
-                        label_overrides = {},
-                    } = assayData[category];
-
-                    // Override category title if label override exists
-                    const categoryTitle = label_overrides?.[title] ?? title;
-                    const categoryAssays = values;
-
-                    return (
-                        <div key={i} className="tissue-category">
-                            <div className="header-container">
-                                <h3>{categoryTitle}</h3>
-                            </div>
-                            <div className="tissue-list-container">
-                                {categoryAssays.length > 0 ? (
-                                    <ul>
-                                        {categoryAssays
-                                            .sort((a, b) => a.localeCompare(b))
-                                            .map((assay, j) => {
-                                                // Override assay title if label override exists
-                                                const assayTitle =
-                                                    label_overrides?.[assay] ??
-                                                    assay;
-
-                                                // Create a link to search for files with this assay
-                                                // Note: The assay link uses the original assay term
-                                                return (
-                                                    <li key={j}>
-                                                        <span>
-                                                            <a
-                                                                href={`/browse/?type=File&${BROWSE_STATUS_FILTERS}&dataset!=No+value&donors.display_title=${itemDetails.display_title}&assays.display_title=${assay}`}
-                                                                target="_blank"
-                                                                rel="noreferrer noopener">
-                                                                {assayTitle}
-                                                            </a>
-                                                        </span>
-                                                    </li>
-                                                );
-                                            })}
-                                    </ul>
-                                ) : (
-                                    <span className="text-secondary">N/A</span>
-                                )}
-                            </div>
-                        </div>
-                    );
-                })}
-            </div>
-        </div>
-    ) : (
-        <div className="detail-content">
-            <div className="detail-header">
-                <i className="icon icon-spin icon-circle-notch"></i>
-            </div>
-        </div>
-    );
-});
-
-// Detail Pane
-const customRenderDetailPane = (
-    itemDetails = {},
-    rowIndex,
-    panelWidth,
-    panelDetails = {}
-) => {
-    // set the corresponding detail pane based on `detailPaneType`
-    let detailPane;
-
-    switch (panelDetails?.detailPaneType) {
-        case 'tissue':
-            detailPane = (
-                <TissueDetailPane
-                    itemDetails={itemDetails}
-                    panelDetails={panelDetails}
-                />
-            );
-            break;
-        case 'assay':
-            detailPane = (
-                <AssayDetailPane
-                    itemDetails={itemDetails}
-                    panelDetails={panelDetails}
-                />
-            );
-            break;
-        default:
-            break;
-    }
-
-    return <div className="detail-pane">{detailPane}</div>;
-};
-
-// A column extension map specifically for browse view file tables.
-export function createBrowseDonorColumnExtensionMap({
-    selectedItems,
-    onSelectItem,
-    onResetSelectedItems,
-}) {
-    const columnExtensionMap = {
-        ...originalColExtMap, // Pull in defaults for all tables
-        // Then overwrite or add onto the ones that already are there:
-        // Select all button
-        '@type': {
-            colTitle: (
-                // Context now passed in from HeadersRowColumn (for file count)
-                <SelectAllFilesButton
-                    {...{ selectedItems, onSelectItem, onResetSelectedItems }}
-                    type="checkbox"
-                />
-            ),
-            hideTooltip: true,
-            noSort: true,
-            widthMap: { lg: 60, md: 60, sm: 60 },
-            render: (result, parentProps) => {
-                return (
-                    <SelectionItemCheckbox
-                        {...{ selectedItems, onSelectItem, result }}
-                        isMultiSelect={true}
-                    />
-                );
-            },
-        },
-        // External ID
-        external_id: {
-            widthMap: { lg: 120, md: 120, sm: 120 },
-            colTitle: <>Donor</>,
-            render: function (result, parentProps) {
-                const {
-                    '@id': atId,
-                    display_title,
-                    external_id,
-                } = result || {};
-
-                return (
-                    <span className="value text-start">
-                        <a
-                            href={atId}
-                            target="_blank"
-                            rel="noreferrer noopener">
-                            {external_id || display_title}
-                        </a>
-                    </span>
-                );
-            },
-        },
-        // Age
-        age: {
-            widthMap: { lg: 100, md: 100, sm: 100 },
-            colTitle: (
-                <span>
-                    Age
-                    <i
-                        className="icon icon-fw icon-info-circle fas"
-                        data-tip="Note: ages 89 and above are denoted as 89+"
-                    />
-                </span>
-            ),
-            render: function (result, parentProps) {
-                if (!result?.age) return null;
-                else {
-                    const ageString = result.age === 89 ? '89+' : result?.age;
-                    return (
-                        <span className="value text-center w-100">
-                            {ageString}
-                        </span>
-                    );
-                }
-            },
-        },
-        // Sex
-        sex: {
-            widthMap: { lg: 80, md: 80, sm: 80 },
-            render: function (result, parentProps) {
-                return (
-                    <span className="value text-center w-100">
-                        {result?.sex?.substring(0, 1) ?? null}
-                    </span>
-                );
-            },
-        },
-        // Tissues
-        tissues: {
-            noSort: true,
-            colAlignment: 'text-end',
-            widthMap: { lg: 120, md: 120, sm: 120 },
-            render: function (result, parentProps) {
-                const {
-                    href,
-                    context,
-                    rowNumber,
-                    detailOpen,
-                    toggleDetailOpen,
-                    detailPaneType,
-                    handleCellClick,
-                } = parentProps;
-
-                const { data, loading } = parentProps?.fetchedProps;
-
-                const tissueFacet = data?.find(
-                    (f) => f.field === 'sample_summary.tissues'
-                );
-                const tissueTerms = tissueFacet?.has_group_by
-                    ? tissueFacet?.original_terms || tissueFacet?.terms
-                    : tissueFacet?.terms;
-                const tissueCount = tissueTerms?.length;
-
-                if (loading) {
-                    return <span className="value text-center loading"></span>;
-                } else {
-                    return tissueCount ? (
-                        <div
-                            className={
-                                'inner-value-container' +
-                                (detailOpen ? ' detail-open' : ' detail-closed')
-                            }>
-                            <CustomTableRowToggleOpenButton
-                                {...{
-                                    result,
-                                    href,
-                                    context,
-                                    rowNumber,
-                                    detailOpen,
-                                    toggleDetailOpen,
-                                    isActive: detailPaneType === 'tissue',
-                                    customToggleDetailClose: (props) => {
-                                        if (detailPaneType === 'assay') {
-                                            handleCellClick('tissue');
-                                        } else {
-                                            handleCellClick(null);
-                                            toggleDetailOpen();
-                                        }
-                                    },
-                                    customToggleDetailOpen: (props) => {
-                                        handleCellClick('tissue');
-                                        toggleDetailOpen();
-                                    },
-                                    toggleOpenIcon: (
-                                        <>
-                                            {tissueCount} Tissue
-                                            {tissueCount > 1 ? 's' : ''}
-                                            <i className="icon icon-circle-plus"></i>
-                                        </>
-                                    ),
-                                    toggleCloseIcon: (
-                                        <>
-                                            {tissueCount} Tissue
-                                            {tissueCount > 1 ? 's' : ''}
-                                            <i className="icon icon-circle-minus"></i>
-                                        </>
-                                    ),
-                                }}>
-                                {tissueCount} Tissue
-                                {tissueCount > 1 ? 's' : ''}
-                            </CustomTableRowToggleOpenButton>
-                        </div>
-                    ) : null;
-                }
-            },
-        },
-        // Assays
-        assays: {
-            noSort: true,
-            colAlignment: 'text-end',
-            widthMap: { lg: 120, md: 120, sm: 120 },
-            render: function (result, parentProps) {
-                const {
-                    href,
-                    context,
-                    rowNumber,
-                    detailOpen,
-                    toggleDetailOpen,
-                    detailPaneType,
-                    handleCellClick,
-                } = parentProps;
-
-                const { data, loading, error } = parentProps?.fetchedProps;
-
-                const assayCount = data
-                    ?.find((f) => f.field === 'assays.display_title')
-                    ?.terms?.reduce(
-                        (acc, curr) => acc + (curr?.terms?.length ?? 1),
-                        0
-                    );
-
-                if (loading) {
-                    return <span className="value text-center loading"></span>;
-                } else {
-                    return assayCount ? (
-                        <div
-                            className={
-                                'inner-value-container' +
-                                (detailOpen ? ' detail-open' : ' detail-closed')
-                            }>
-                            <CustomTableRowToggleOpenButton
-                                {...{
-                                    result,
-                                    href,
-                                    context,
-                                    rowNumber,
-                                    detailOpen,
-                                    toggleDetailOpen,
-                                    isActive: detailPaneType === 'assay',
-                                    customToggleDetailClose: (props) => {
-                                        if (detailPaneType === 'tissue') {
-                                            handleCellClick('assay');
-                                        } else {
-                                            handleCellClick(null);
-                                            toggleDetailOpen();
-                                        }
-                                    },
-                                    customToggleDetailOpen: (props) => {
-                                        handleCellClick('assay');
-                                        toggleDetailOpen();
-                                    },
-                                    toggleOpenIcon: (
-                                        <>
-                                            {assayCount} Assay
-                                            {assayCount > 1 ? 's' : ''}
-                                            <i className="icon icon-circle-plus"></i>
-                                        </>
-                                    ),
-                                    toggleCloseIcon: (
-                                        <>
-                                            {assayCount} Assay
-                                            {assayCount > 1 ? 's' : ''}
-                                            <i className="icon icon-circle-minus"></i>
-                                        </>
-                                    ),
-                                }}></CustomTableRowToggleOpenButton>
-                        </div>
-                    ) : null;
-                }
-            },
-        },
-        // Files
-        files: {
-            noSort: true,
-            colAlignment: 'text-end',
-            widthMap: { lg: 90, md: 90, sm: 90 },
-            render: function (result, parentProps) {
-                const { data, loading, error } = parentProps?.fetchedProps;
-
-                const fileCount = data
-                    ?.find((f) => f.field === 'type')
-                    ?.terms?.find((term) => term.key === 'File')?.doc_count;
-
-                if (loading) {
-                    return (
-                        <span className="value text-center loading">
-                            {/* <i className="icon icon-circle-notch icon-spin fas"></i> */}
-                        </span>
-                    );
-                } else {
-                    return fileCount ? (
-                        <a
-                            className="value text-center"
-                            href={`/browse/?type=File&${BROWSE_STATUS_FILTERS}&dataset!=No+value&donors.display_title=${result?.display_title}`}>
-                            {fileCount} File{fileCount > 1 ? 's' : ''}
-                        </a>
-                    ) : null;
-                }
-            },
-        },
-        file_size: {
-            noSort: true,
-            colAlignment: 'text-end',
-            widthMap: { lg: 90, md: 90, sm: 90 },
-            render: function (result, parentProps) {
-                const {
-                    href,
-                    context,
-                    rowNumber,
-                    detailOpen,
-                    toggleDetailOpen,
-                } = parentProps;
-
-                const { data, loading, error } = parentProps?.fetchedProps;
-
-                const fileSize = data?.find(
-                    (f) => f.field === 'file_size'
-                )?.sum;
-
-                if (loading) {
-                    return (
-                        <span className="value text-center loading">
-                            {/* <i className="icon icon-circle-notch icon-spin fas"></i> */}
-                        </span>
-                    );
-                } else {
-                    return fileSize ? (
-                        <span className="value text-center">
-                            {valueTransforms.bytesToLargerUnit(
-                                fileSize,
-                                0,
-                                false,
-                                true
-                            )}{' '}
-                            {valueTransforms.bytesToLargerUnit(
-                                fileSize,
-                                0,
-                                true,
-                                false
-                            )}
-                        </span>
-                    ) : null;
-                }
-            },
-        },
-        // Hardy Scale
-        hardy_scale: {
-            widthMap: { lg: 140, md: 140, sm: 140 },
-            render: function (result, parentProps) {
-                return (
-                    <span className="value text-center">
-                        {result?.hardy_scale ?? null}
-                    </span>
-                );
-            },
-        },
-        // Data Type
-        data_type: {
-            noSort: true,
-            widthMap: { lg: 124, md: 124, sm: 124 },
-            render: function (result, parentProps) {
-                const { data_type = [] } = result || {};
-                if (data_type.length === 0) {
-                    return null;
-                } else if (data_type.length === 1) {
-                    return data_type[0];
-                } else {
-                    return data_type.join(', ');
-                }
-            },
-        },
-        // Released
-        'file_status_tracking.release_dates.initial_release_date': {
-            colTitle: 'Released',
-            widthMap: { lg: 115, md: 115, sm: 115 },
-            render: function (result, parentProps) {
-                const value =
-                    result?.file_status_tracking?.release_dates
-                        ?.initial_release_date;
-                if (!value) return null;
-                return <span className="value text-end">{value}</span>;
-            },
-        },
-        // Platform
-        'sequencers.display_title': {
-            widthMap: { lg: 170, md: 160, sm: 150 },
-        },
-        // Format
-        'file_format.display_title': {
-            widthMap: { lg: 100, md: 90, sm: 80 },
-        },
-        // Software
-        'software.display_title': {
-            widthMap: { lg: 151, md: 151, sm: 151 },
-        },
-        // Date Created
-        date_created: {
-            widthMap: { lg: 151, md: 151, sm: 151 },
-            default_hidden: true,
-            render: function (result, parentProps) {
-                const value = result?.date_created;
-                if (!value) return null;
-                return (
-                    <span className="value text-end">
-                        <LocalizedTime
-                            timestamp={value}
-                            formatType="date-file"
-                        />
-                    </span>
-                );
-            },
-        },
+export function createBrowseDonorColumnExtensionMap(props) {
+    const { columnExtensionMap, columns, hideFacets } =
+        createBaseDonorColumnExtensionMap(props);
+    return {
+        columnExtensionMap,
+        columns,
+        hideFacets: [...hideFacets, 'tissues'],
     };
-
-    const columns = {
-        '@type': {
-            title: 'Selected',
-        },
-        external_id: {
-            title: 'Donor',
-        },
-        age: {
-            title: 'Age',
-        },
-        sex: {
-            title: 'Sex',
-        },
-        tissues: {
-            title: 'Tissues',
-        },
-        assays: {
-            title: 'Assays',
-        },
-        files: {
-            title: 'Files',
-        },
-        file_size: {
-            title: 'File Size',
-        },
-        hardy_scale: {
-            title: 'Hardy Scale',
-        },
-    };
-
-    const hideFacets = [
-        'tissues',
-        'dataset',
-        'file_sets.libraries.analytes.samples.sample_sources.code',
-        'status',
-        'validation_errors.name',
-        'version',
-        'study',
-        'submission_centers.display_title',
-        'software.display_title',
-        'tags',
-    ];
-
-    return { columnExtensionMap, columns, hideFacets };
 }
 
 // Search Table
@@ -712,14 +51,12 @@ const BrowseDonorSearchTable = (props) => {
         onResetSelectedItems, // From SelectedItemsController
     };
 
-    // Pass modified context to CommonSearchView to set default filters
     const passProps = {
         ...props,
         context: {
             ...context,
             clear_filters: BROWSE_LINKS.donor,
         },
-        // Provide a search for populating custom column(s)
         customColumnSearchHref: (result) =>
             `/peek-metadata/?additional_facet=file_size&${BROWSE_STATUS_FILTERS}&dataset!=No+value&type=File&donors.display_title=` +
             result?.display_title,
@@ -740,7 +77,6 @@ const BrowseDonorSearchTable = (props) => {
     const { columnExtensionMap, columns, hideFacets } =
         createBrowseDonorColumnExtensionMap(selectedFileProps);
 
-    // Custom sort functions for specific facet lists
     const facetListSortFxns = {
         hardy_scale: (a, b) => {
             return a.key - b.key;
